utils.py
# Import the necessary libraries and frameworks
import cv2 as cv
import numpy as np
import pickle
import h5py
import imageio
from skimage.metrics import structural_similarity as ssim
from sklearn.metrics import mean_squared_error

# Define the function that loads the input data from a path
def load_input_data(input_path):
    # Initialize the input data list
    input_data = []

    # Check the file extension of the input path
    if input_path.endswith(".jpg") or input_path.endswith(".png"):
        # Load the image from the input path
        image = cv.imread(input_path)

        # Append the image to the input data list
        input_data.append(image)

    elif input_path.endswith(".mp4") or input_path.endswith(".mov"):
        # Load the video from the input path
        video = cv.VideoCapture(input_path)

        # Loop through the frames of the video
        while video.isOpened():
            # Read the frame from the video
            ret, frame = video.read()

            # Check if the frame is valid
            if ret:
                # Append the frame to the input data list
                input_data.append(frame)
            else:
                # Break the loop
                break

        # Release the video
        video.release()

    # Return the input data list
    return input_data

# Define the function that preprocesses the input data, such as resizing, cropping, or normalizing
def preprocess_input_data(input_data):
    # Initialize the output data list
    output_data = []

    # Loop through the input data
    for data in input_data:
        # Resize the data to a fixed size
        data = cv.resize(data, (800, 600))

        # Crop the data to a square shape
        data = data[0:600, 100:700]

        # Normalize the data to a range of [0, 1]
        data = data / 255.0

        # Append the data to the output data list
        output_data.append(data)

    # Return the output data list
    return output_data

# Define the function that saves the output data to a path
def save_output_data(output_data, output_path):
    # Check the file extension of the output path
    if output_path.endswith(".pkl"):
        # Save the output data as a pickle file
        with open(output_path, "wb") as f:
            pickle.dump(output_data, f)

    elif output_path.endswith(".h5"):
        # Save the output data as a HDF5 file
        with h5py.File(output_path, "w") as f:
            f.create_dataset("output_data", data=output_data)

    elif output_path.endswith(".obj"):
        # Save the output data as a 3D model file
        with open(output_path, "w") as f:
            for data in output_data:
                f.write("v {} {} {}\n".format(data[0], data[1], data[2]))

    elif output_path.endswith(".mp4") or output_path.endswith(".mov"):
        # Save the output data as a MP4 or MOV file
        writer = cv.VideoWriter(output_path, cv.VideoWriter_fourcc(*"MP4V"), 30, (800, 600))
        for data in output_data:
            writer.write(data)
        writer.release()

    elif output_path.endswith(".gif") or output_path.endswith(".png"):
        # Save the output data as a GIF or PNG file
        imageio.mimwrite(output_path, output_data)

# Define the function that evaluates the output data, using metrics such as mean squared error, structural similarity index, or frames per second
def evaluate_output_data(output_data, input_data):
    # Ensure input_data and output_data are numpy arrays
    input_data = np.array(input_data)
    output_data = np.array(output_data)

    # Compute MSE
    mse = mean_squared_error(input_data.flatten(), output_data.flatten())

    # Compute SSIM
    ssim_value = ssim(input_data, output_data, multichannel=True)

    # Compute FPS (assuming output_data is a list of frames)
    fps = len(output_data) / 30  # Assuming 30 seconds of video

    # Print the metrics
    print(f"Mean Squared Error: {mse:.4f}")
    print(f"Structural Similarity Index: {ssim_value:.4f}")
    print(f"Frames Per Second: {fps:.4f}")

    return mse, ssim_value, fps
